US7 Done and automation test passed
